Feature/user toko

+++ backend/src/main/java/com/daily_app/demo/Repository/UserRepository.java
@@ -1,4 +1,5 @@
-package com.daily_app.demo.Repository;
+//　松原編集
+package com.daily_app.demo.Repository; 
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
@@ -6,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import com.daily_app.demo.Entity.User;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<User, Integer> {
     
 }
